backend/requirements.txt
aioredis==2.0.1
alembic==1.16.1
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
bcrypt==4.0.0
better-profanity==0.7.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.0
coverage==7.11.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.11.0
fastapi==0.115.12
fastapi-limiter==0.1.6
freezegun==1.5.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.4
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==6.2.0
requests==2.32.5
requests-oauthlib==2.0.0
responses==0.25.8
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.42.0
starlette==0.46.2
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
